moved recipeform to pages and got rid of dead code
