+++ frontend/src/main.tsx
@@ -1,7 +1,6 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-// import RecipeForm from './RecipeForm.tsx'
 import App from './App'
 
 createRoot(document.getElementById('root')!).render(
